+++ src/dune_rules/opam_create.ml
@@ -48,7 +48,7 @@ let default_build_command =
          (Lexbuf.from_string ~fname:"<internal>"
             {|
 [
-  [ "dune" "subst" "--root" "." ] {dev}
+  [ "dune" "subst" ] {dev}
   [ "dune" "build" "-p" name "-j" jobs "--promote-install-files" "false"
       "@install"
       "@runtest" {with-test}