+++ src/dune_filesystem_stubs/readdir.c
@@ -18,7 +18,8 @@ typedef struct dirent directory_entry;
 value val_file_type(int typ) {
   switch(typ)
     {
-    case DT_REG:
+#ifndef __HAIKU__     
+   case DT_REG:
       return Val_int(0);
     case DT_DIR:
       return Val_int(1);
@@ -34,6 +35,7 @@ value val_file_type(int typ) {
       return Val_int(6);
     case DT_UNKNOWN:
       return Val_int(7);
+#endif
     default:
       return Val_int(7);
     }
@@ -57,7 +59,11 @@ CAMLprim value caml__dune_filesystem_stubs__readdir(value vd)
   v_filename = caml_copy_string(e->d_name);
   v_tuple = caml_alloc_small(2, 0);
   Field(v_tuple, 0) = v_filename;
+#ifndef __HAIKU__
   Field(v_tuple, 1) = val_file_type(e->d_type);
+#else
+  Field(v_tuple, 1) = Val_int(7);
+#endif
   CAMLreturn(v_tuple);
 }