+++ CHANGES.md
@@ -1,6 +1,8 @@
 2.9.1 (unreleased)
 ------------------
 
+- Don't use `subst --root` in Opam files (#4806, @MisterDA)
+
 - Fix compilation on Haiku (#4885, @Sylvain78)
 
 2.9.0 (29/06/2021)