+++ CHANGES.md
@@ -1,6 +1,13 @@
-unreleased - 2.9 branch
------------------------
+2.9.1 (unreleased)
+------------------
 
+- Don't use `subst --root` in Opam files (#4806, @MisterDA)
+
+- Fix compilation on Haiku (#4885, @Sylvain78)
+
+- Allow depending on `ocamldoc` library when `ocamlfind` is not installed.
+  (#4811, fixes #4809, @nojb)
+  
 - Fix `(enabled_if ...)` for installed libraries (#4824, fixes #4821, @dra27)
 
 2.9.0 (29/06/2021)